--- a/Package.swift
+++ b/Package.swift
@@ -41,6 +41,7 @@ let package = Package(
             dependencies: ["XcbeautifyLib"],
             resources: [
                 .copy("ParsingTests/clean_build_xcode_15_1.txt"),
+                .copy("ParsingTests/large_xcodebuild_log.txt"),
             ]
         ),
     ]