--- a/Tests/XcbeautifyLibTests/ParsingTests/ParsingTests.swift
+++ b/Tests/XcbeautifyLibTests/ParsingTests/ParsingTests.swift
@@ -38,4 +38,41 @@ final class ParsingTests: XCTestCase {
         // There's a regression whenever `uncapturedOutput` is greater than the current magic number.
         XCTAssertEqual(uncapturedOutput, 2218)
     }
+
+    func testLargeXcodebuildLog() throws {
+        let url = Bundle.module.url(forResource: "large_xcodebuild_log", withExtension: "txt")!
+
+        var buildLog: [String] = try String(contentsOf: url)
+            .components(separatedBy: .newlines)
+
+        let parser = Parser(
+            colored: false,
+            renderer: .terminal,
+            preserveUnbeautifiedLines: false,
+            additionalLines: {
+                guard !buildLog.isEmpty else {
+                    XCTFail("The build log should never be empty when fetching additional lines.")
+                    return nil
+                }
+                return buildLog.removeFirst()
+            }
+        )
+
+        var uncapturedOutput = 0
+
+        while !buildLog.isEmpty {
+            let line = buildLog.removeFirst()
+            if !line.isEmpty, parser.parse(line: line) == nil {
+                uncapturedOutput += 1
+            }
+        }
+
+        // The following is a magic number.
+        // It represents the number of lines that aren't captured by the Parser.
+        // Slowly, this value should decrease until it reaches 0.
+        // It uses `XCTAssertEqual` instead of `XCTAssertLessThanOrEqual` as a reminder.
+        // Update this magic number whenever `uncapturedOutput` is less than the current magic number.
+        // There's a regression whenever `uncapturedOutput` is greater than the current magic number.
+        XCTAssertEqual(uncapturedOutput, 77104)
+    }
 }