## Changes

Drastically improve xcbeautify's performance (time and memory). After testing, this change should drop xcbeautify's execution time by ~90% and memory footprint by ~99%.

- Reduce unnecessary type instantiations, namely `NSRegularExpression`.
- Skip parsing empty lines.
- Introduce time-tracking logic and large xcodebuild log test to (1) find and fix the existing issue and (2) prevent future performance regressions. 

## Testing

Average results after formatting **large_xcodebuild_log.txt** (99.9 MB xcodebuild log).

| Measurement    | Before | After | Absolute Difference | Percentage Difference |
| -------- | ------- | ------- | ------- | ------- |
| Memory  | 2.21 GB  | 8.3 MB | -2254.74 MB / -2.2 GB | -99.63% |
| Time | 66 seconds | 5.9 seconds | -60.1 seconds | -91% |